+++ src/harlequin/cli.py
@@ -1,17 +1,25 @@
 from __future__ import annotations
 
 import sys
-from typing import Any, Callable
+from pathlib import Path
+from typing import Any, Callable, Sequence
 
 import rich_click as click
 
 from harlequin import Harlequin
 from harlequin.adapter import HarlequinAdapter
+from harlequin.config import get_config_for_profile
+from harlequin.exception import HarlequinConfigError, pretty_print_error
 
 if sys.version_info < (3, 10):
     from importlib_metadata import entry_points, version
 else:
     from importlib.metadata import entry_points, version
+
+# configure defaults
+DEFAULT_ADAPTER = "duckdb"
+DEFAULT_LIMIT = 100_000
+DEFAULT_THEME = "monokai"
 
 # configure the rich click interface (mostly --help options)
 DOCS_URL = "https://harlequin.sh/docs/getting-started"
@@ -53,7 +61,15 @@ click.rich_click.OPTION_GROUPS = {
     "harlequin": [
         {
             "name": "Harlequin Options",
-            "options": ["--adapter", "--theme", "--limit", "--version", "--help"],
+            "options": [
+                "--profile",
+                "--config-path",
+                "--adapter",
+                "--theme",
+                "--limit",
+                "--version",
+                "--help",
+            ],
         },
     ]
 }
@@ -108,9 +124,36 @@ def build_cli() -> click.Command:
         nargs=-1,
     )
     @click.option(
+        "-P",
+        "--profile",
+        help=(
+            "Select a profile from an available config file to load its values. "
+            "Other options passed here will take precedence over those loaded "
+            "from the profile. Use the special profile named None to use Harlequin's "
+            "defaults, instead of the default profile specified in the config "
+            "file."
+        ),
+    )
+    @click.option(
+        "--config-path",
+        help=(
+            "By default, Harlequin finds files named .harlequin.toml in the "
+            "current directory and the home directory (~) and merges them. "
+            "Use this option to specify the full path to a config file at "
+            "a different location."
+        ),
+        type=click.Path(
+            exists=True,
+            file_okay=True,
+            dir_okay=False,
+            resolve_path=True,
+            path_type=Path,
+        ),
+    )
+    @click.option(
         "-t",
         "--theme",
-        default="monokai",
+        default=DEFAULT_THEME,
         show_default=True,
         help=(
             "Set the theme (colors) of the Harlequin IDE. "
@@ -121,7 +164,7 @@ def build_cli() -> click.Command:
     @click.option(
         "--limit",
         "-l",
-        default=100_000,
+        default=DEFAULT_LIMIT,
         type=click.IntRange(min=0),
         help=(
             "Set the maximum number of rows that can be loaded into Harlequin's "
@@ -131,7 +174,7 @@ def build_cli() -> click.Command:
     @click.option(
         "-a",
         "--adapter",
-        default="duckdb",
+        default=DEFAULT_ADAPTER,
         show_default=True,
         type=click.Choice(list(adapters.keys()), case_sensitive=False),
         help=(
@@ -142,10 +185,8 @@ def build_cli() -> click.Command:
     @click.pass_context
     def inner_cli(
         ctx: click.Context,
-        conn_str: tuple[str],
-        theme: str,
-        limit: int,
-        adapter: str,
+        profile: str | None,
+        config_path: Path | None,
         **kwargs: Any,
     ) -> None:
         """
@@ -155,6 +196,13 @@ def build_cli() -> click.Command:
         connection strings (or paths to local db files) for databases to open with
         Harlequin.[/]
         """
+        # load config from any config files
+        try:
+            config = get_config_for_profile(config_path=config_path, profile=profile)
+        except HarlequinConfigError as e:
+            pretty_print_error(e)
+            ctx.exit(2)
+
         # prune the kwargs to only those that don't have their default arguments
         params = list(kwargs.keys())
         for k in params:
@@ -163,15 +211,25 @@ def build_cli() -> click.Command:
                 == click.core.ParameterSource.DEFAULT  # type: ignore
             ):
                 kwargs.pop(k)
+            # conn_str is an arg, not an option, so get_paramter_source is always CLI
+            elif k == "conn_str" and kwargs[k] == tuple():
+                kwargs.pop(k)
+
+        # merge the config and the cli options
+        config.update(kwargs)
 
         # load and instantiate the adapter
-        adapter_cls: type[HarlequinAdapter] = adapters[adapter]
-        adapter_instance = adapter_cls(conn_str=conn_str, **kwargs)
+        adapter = config.pop("adapter", DEFAULT_ADAPTER)
+        conn_str: Sequence[str] = config.pop("conn_str", tuple())  # type: ignore
+        if isinstance(conn_str, str):
+            conn_str = (conn_str,)
+        adapter_cls: type[HarlequinAdapter] = adapters[adapter]  # type: ignore
+        adapter_instance = adapter_cls(conn_str=conn_str, **config)
 
         tui = Harlequin(
             adapter=adapter_instance,
-            max_results=limit,
-            theme=theme,
+            max_results=config.get("limit", DEFAULT_LIMIT),  # type: ignore
+            theme=config.get("theme", DEFAULT_THEME),  # type: ignore
         )
         tui.run()